pyproject.toml
[tool.poetry]
name = "ecommerce-admin-backendd"
version = "0.1.0"
description = "백어드민 백엔드"
authors = ["창민 <ckd29672@naver.com>"]
readme = "README.md"
packages = [{include = "ecommerce_admin_backendd"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
sqlalchemy = "^2.0.4"
pymysql = "^1.0.2"
dependency-injector = "^4.41.0"
alembic = "^1.9.4"
hackle-sdk = "^3.3.0"
slack-sdk = "^3.20.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
black = "^23.1.0"
isort = "^5.12.0"
mypy = "^1.0.1"
httpx = "^0.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
